+++ LCD_digits/test/display.js
@@ -1,11 +1,11 @@
 let chai = require('chai');
 let should = chai.should();
-let calc = require('../calculator');
+let displayer = require('../displayer');
 
 describe('Calculator Functions', () => {
     describe('sum()', () => {
         it('should return the sum of two numbers', () => {
-            let res = calc.sum(5,3);
+            let res = displayer.sum(5,3);
             res.should.be.equal(8);
         });
     });
